fetch the weather when the network is available; basically i was sick of having a blank wallpaper for 30 minutes after log in just because network manager hadn't connected in the time it takes plasma to get on its feet.
this is such a simple pattern, yet one that is pretty common (getting stuff from the network at regular intervals) should it be put into DataEngine, perhaps protected: setRequiresNetwork(bool)? or perhaps a few different levels such as "NoNetworking (presumably the default), NetworkOptional, NetworkRequired" so that we can at some point turn it on/off depending on whether or not the networking is expensive/cheap/etc? none/optional/required is perhaps not fine grained enough? discuss .... :)

CCMAIL:plasma-devel@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=954107