Fix layout of Fonts KCM
My original assumption regarding the zero-spaced font selectors
was that this was intended to cram more fonts into the dialog.

But code inspection did not reveal the expected setSpacing(0) call.
The reason for this crammed layout were calls to QLayout::addItem()
instead of QLayout::addLayout() when adding child layouts.

This may or may not be a Qt bug, but looking at Qt documentation,
the addItem() "function is not usually called in application code".


svn path=/trunk/KDE/kdebase/workspace/; revision=959331