* create a TimeSource that parses out the name of the source just once instead of on every update and which handle the sun/moon things internall instead of having that scattered all over the place * only update the sunrise, sunset and moon phase info once a day no matter how often called * initialize the time_t time variable correctly in the moon phase calculation * use a namespace instead of a class for the solar positioning code (it was all static members)
in the general case, this doesn't actually provide any real speed performance benefits as the worst case scanarios weren't actually that bad and probably never actually get called from our code. this is sill a lot tidier (and a few less bugs to boot ;) and there are significantly fewer memory allocations that hapen.

svn path=/trunk/KDE/kdebase/workspace/; revision=965474