try to fix up the shutdown stuff
first of all, deprecate the braindead ShutdownTypeLogout again and alias
it to ShutdownTypeNone internally, as the latter means "logout only"
(that enum refers to the machine's state, not the session's).

unify the semantics of ksmserver's d-bus interface to interpret a
shutdown request with a pre-selected shutdown type as needing only a
confirmation, not a full choice. this is a behavior break from the
original implementation, but that actually happened in 4.2 already - i'm
just fixing it.

all "generic" logout actions (alt-ctrl-del, lock_logout applet, plasma
desktop menu) now always pop up the full shutdown dialog, irrespective
of the "confirm shutdown" option.
"specialized" logout actions (like in kickoff's "leave" menu) continue
to respect the "confirm shutdown" option. if the confirmation is
enabled, the reduced shutdown dialog will appear as before (only some
implementation details changed).

BUG: 152110
BUG: 177139
BUG: 182466

svn path=/trunk/KDE/kdebase/workspace/; revision=976028