showing the remaining time is now opt-in for geeks
- Only show the time in the popup when m_showRemainingTime is true
- m_showRemainingTime can still be set manually in the config file
- Removed options from the config UI

I've disabled this feature for now since it doesn't work well enough
in a lot of cases. For those that really can't live without that
information, you can still enable it in the config, add a line such
as:

showRemainingTime=false

in the applet's configuration, and it comes back (as unreliable as
this information actually is).

CCMAIL:plasma-devel@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=978089