This appears to have survived review on kde-core-devel, so commit.
This changes the default of MALLOC_CHECK_ (glibc memory error checking)
from 2 (abort with no error message) to 3 (abort with error message)
if glibc >= 2.9.  Prior glibc versions supported 3 but always printed
a debugging blurb which would break systems.

Slight behavior modification is that MALLOC_CHECK_ is not only set if it
was not already set in the environment, and only if /lib/libc.so.6 is
detected.

svn path=/trunk/KDE/kdebase/workspace/; revision=991495