Backport of r991543: Add the keystate engine to the build which I forgot when moving it from kdereview.
While the engine currently doesn't feature a Plasmoid using it, the plan is to release one based on QtScript separately during the 4.3 release cycle. As
an applet displaying the state of the modifier keys is still badly missing we decided it was important to ship this dataengine.

I'm sorry for the inconveniences this causes - especially for the translation teams.

CCMAIL:release-team@kde.org
CCMAIL:kde-i18n-doc@kde.org
CCMAIL:aseigo@kde.org

svn path=/branches/KDE/4.3/kdebase/workspace/; revision=991554