color kcm: Remove focus/hover background preview
While it's not "illegal" to use background-tinted-to-decoration, there is no requirement that this color contrast with the unmodified decoration color. Therefore, change the set preview to not imply that this is the case. (Technically, it doesn't even need to contrast well enough for legible text.)

Noticed thanks to an attachment in the referenced bug, and, while this doesn't fix said bug, it makes it slightly less broken due to one fewer item in the long row.

CCBUG: 198978

svn path=/trunk/KDE/kdebase/workspace/; revision=996037