Allow adding submenus of the application launcher to the panel or the desktop by right clicking on them in the main menu.
This works by storing the relative path and icon name of KServiceGroups
(== submenus) in the application model used by the menu

When an 'Add to Panel/Desktop' context menu action is triggered, add a
simplelauncher applet (instead of an Icon).

This receives the relative path of the submenu and the icon name in its
args, and when it builds its menu, it walks the applications model and
locates the submenu, rooting the menu at that index rather than the real
root, and setting the icon for the launcher to that of the submenu.

When the applet is showing a submenu, the view types and number of
recent applications to show are not shown in the config UI.

TODO: set tooltip to title of submenu instead of 'Application Launcher
Menu'.

BUG: 189583


svn path=/trunk/KDE/kdebase/workspace/; revision=1015064