Forwardport r801043 from 3.5 branch, it somehow got lost.
"While the internal buffer /proc/cpuinfo is read into is 32k is size, reading
these 32k in a single chunk is not possible: Only up to a page wort of
information is being returned in a single read (reading stops as soon as the
full block for a subsequent CPU wouldn't fit into the buffer space remaining in
the page anymore). The effect is that only 5 or 6 CPUs (depending on the exact
size info for one CPU has) get their frequency displayed (and the respective
sensor offered for selection) on x86/x86-64 (where the page size is 4k).
Consequently, multiple reads (until zero bytes are read) must be issued to
obtain all information."

svn path=/trunk/KDE/kdebase/workspace/; revision=1015584