Keep in sync with recent font size changes
Not sure if/where the font size changes are discussed, but I think
making the default size smaller is dangerous; I am thinking about
a 15" screen with only 800x600 resolution, but am too lazy right
now to do the math...

Since I am not in the position to decide, I will commit this in
order to keep the changes in sync with r1049144.

If there is a need to make further changes to kglobalsettings.cpp,
please also change this file.

CCMAIL: kde-core-devel@kde.org
CCMAIL: kde-artists@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=1051172