So far one rootGroup was used as library representation of the visualization. For manual grouping/sorting the tasks were remembered to restore the old grouping/sorting after a desktop change.
From now on we use another approach, by creating a rootGroup for each desktop.
This allowed to remove all the code which was used to store the grouping/sorting (which was also rather messy IMHO), since no reevaluation is done when we return to the old desktop.
The only change to the library user is:
Instead that all tasks are removed when the desktop changes, the rootGroup changes.

Different grouping of tasks on different screens is not supported, since tasks cannot be on several screens at the same time afaik.

svn path=/trunk/KDE/kdebase/workspace/; revision=1053974