Bye-bye deprecated use of Plasma::Animator
We now use QPropertyAnimations for the painting of the pixmaps. This
saves us a good amount of icky animation code, a slew of compiler-
warnings and some householding of the code. Moreover, it fixes

BUG:204725,211001

the final blow. (The label didn't appear in the popup immediately, but
it did in the panel)

svn path=/trunk/KDE/kdebase/workspace/; revision=1054836