Re-enable the hidden config option to show the remaining time in the battery applet
You can enable showing the remaining time in your plasma-desktop-appletsrc (or
plasma-netbook-appletsrc), locate the section for the "battery" plugin and add
"showRemainingTime=true" on a new line. Make sure you're not editing the section
for the battery's extenderitem, but the battery itself.) In my config, this
looks like:

[Containments][3][Applets][7][Configuration][Applets][30]
geometry=140,2.5,30,24
immutability=1
plugin=battery
zvalue=0

[Containments][3][Applets][7][Configuration][Applets][30][Configuration]
Share=false
showBatteryString=false
showRemainingTime=true

(Obviously, the last line is relevant.)

We're not offering this as an option in the GUI since this information is,
for a lot of users not reliable enough. The label for the time remaining
option is disabled for now since I couldn't re-use the existing string.
Should be uncommented once 4.5 opens.

I hope we can all be happy now :)

CCMAIL:plasma-devel@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=1054869