polish up jeff's fix
the fix itself looks like it is just brushing over symptoms, but
kgreeterplugin.h effectively says that redundand gplugStart() calls are
to be expected - even if it was meant for a completely different context.
i'm too lazy now to verify that similar races cannot occurr with the
other callbacks. :D

svn path=/trunk/KDE/kdebase/workspace/; revision=1089705