fix CVE-2010-0436: local root hole relating to command sockets
the user owns the socket directory. if he prevented its deletion (e.g.,
by creating another file in it), he would subsequently get the chance to
inject a symlink to an arbitrary file whose permissions kdm would
conveniently change to 0666.

chmod()ing a directory to the user opens a can of worms, as he might
do all kinds of nasty things with it which would prevent us from
reclaiming it.

so instead rely on the system's ability to honor file ownership and
permissions of the socket's inode instead of its parent directory.

for systems where we cannot rely on this behavior (solaris), we create a
new randomly named socket directory for each session and symlink it from
the proper name. if the user tries to mess with us, he will pollute the
top-level socket directory but cause no real harm.

discovered by Sebastian Krahmer from the SUSE security team.


svn path=/trunk/KDE/kdebase/workspace/; revision=1114416