eliminate small possibility of privacy leak
if the hack was suspended right after an override window popped above us
and before we had a chance to restack and erase, we'd take a snapshot of
that window and later restore it. now we still would take it, but we'd
delete it again if restacking is done soon afterwards. if the snapshot
is restored before we delete it, then we'll just erase the restored
image as if there never was a snapshot involved.

svn path=/trunk/KDE/kdebase/workspace/; revision=1121609