make use of the activity class
oh joy, git lost my merged commit message. anyways, it went something
like this:

create a class to encapsulate the activity stuff.

areas that still need
implementing should all be marked with TODO, except that ActivityIcon
should make use of the Activity's name and thumbnail. the idea is to
keep all the gory implementation inside Activity and PlasmaApp, so that
activitymanager is just for display logic.

it'd be nice if the Activity class n'stuff were exposed through a
dataengine someday, so that people could write other UIs, like maybe a
quick-switcher for the panel.

svn path=/trunk/KDE/kdebase/workspace/; revision=1122460