Show an OSD notification when the laptop brightness keys are pressed and change the brightness if it's not handled by the hardware or the kernel.
Quick implementation overview:
The PowerDevil daemon registers global shortcuts for the
Qt::Key_MonBrightnessUp/Down keys and forwards them to Solid.
Solid then checks if it needs to adjust the brightness.
After that PowerDevil emits the dbus signal brightnessChanged().
The battery applet listens to that signal and displays an OSD
simliar to the one KMix shows.

http://reviewboard.kde.org/r/1942/

BUG: 182400

svn path=/trunk/KDE/kdebase/workspace/; revision=1124692