try a completely different approach to populating the results list:
* maximum 50 items; prevent crazy runners from destroying performance. at some point we may want to institute limits elsewhere and/or paging between the shown and unshown items
* pre-allocate all the items and just set the match data on them

this means a lot fewer list iterations, no hash to populate and lookup and no allocations or scene add/removes during the matching. works well in testing here, hopefully it adds to the fluidity for others.

this currently breaks the scroll buttons, but those will be going tomorrow to be replaced by something a bit nicer.
this commit fixes at least a couple of bugs on b.k.o, but i don't have them handy. will close manually later.

svn path=/trunk/KDE/kdebase/workspace/; revision=1128685