Proper migration code also, make extra certain the current activity is initialized (and make a guess at what to initialize it *to*).
this hasn't been extensively tested, so I'd appreciate it if people
could play with it :) note that upgrading from a beta to this code
(trunk or.. rc2?) may not work as well as upgrading from 4.4. you did a backup
before trying the beta, right? ;) (it won't break anything, it just
won't fix things the beta may have broken)

CCBUG: 243003
CCBUG: 238699
BUG: 242983


svn path=/trunk/KDE/kdebase/workspace/; revision=1145652