First version of blacklist update script. It blacklists for Lanczos Intel: GM45 Express with Mesa 7.8.2 and 965GM with Mesa 7.7.1 and 7.8.2 Ati: DRI R600 with Mesa 7.8.1 and 7.8.2
For Blur
Intel: currently no card (we have to block all cards in the effect)
Ati: Mesa drivers as for Lanczos and Radeon HD 3650 with fglrx driver 3.3.9901
NVIDIA: GeForce 6150 with driver series 195

This is what I could read from the bug reports. There must be more versions which have to be blacklisted.
Please post your glxinfo for your cards to the bug report if it is not yet in the list.
I do not yet backport the blacklist to branch in the hope to get more versions for the list.
BUG: 243181
BUG: 242985

svn path=/trunk/KDE/kdebase/workspace/; revision=1150382