Revert r1157716: control amount of brightness sliders
This patch produces 4 sliders on my machine (thinkpad T60).

As 4 sliders on OK machines are worse than 1 slider We have to  revert it
for now until we find a proper solution for it. I don't see anything I'd
happily commit to the stable branch, so I'll backport this one and we'll
let the bug in until we have something that works *and* doesn't break
machines that are otherwise OK.

CCBUG:199520
CCMAIL:zajec5@gmail.com

svn path=/trunk/KDE/kdebase/workspace/; revision=1158799