Add support to one more text provider to share dataengine
Include privatepaste.com as one of the service providers of the share
data engine. Please note that this implementation does not support private
pastes and if somebody wants to support that, it should be implemented in
a new plugin (and made available through GHNS soon).

I'm still studying how to provide ways of interaction between the plugins
and the main widget to ask questions like "passphrases". for now I would suggest
to base the code on this one and create random passphrases that should be
told to the user using notifications for example (need this support on the
applet).

If someone is interested in implementing this, get in touch with me and I
can help and guide the development of this feature.

BUG: 226367

svn path=/trunk/KDE/kdebase/workspace/; revision=1168991