Forward port rev 1169432: Fix window movment regression (e.g. pack) in 4.5.
The regression was caused in tiling which set the geometry to a different one again.
Nikhil please verify that tiling is still working correctly. Looking at the code this
commit should be side-effect free. When coming from Client::move() moveResizeGeom and
initialMoveResizeGeom may not be set. So I think the code in notifyTilingWindowMove()
cannot work at all ;-)
CCBUG: 241049

svn path=/trunk/KDE/kdebase/workspace/; revision=1169433