Work around Qt 4.6 bug in QComboBox::setMaxVisibleItems
Hopefully this is the last workaround for a workaround for a workaround
of a bug that should not be there in the first place ...

Since this Qt bug is fixed in Qt 4.7 we need the version check. Note that
this is a _compile time_ check, so if you upgrade to Qt 4.7, you have to
recompile Plasma, otherwise it will be limited to just 3 visible entries.

Please test in trunk, so that it can be (finally) backported to 4.5.2.

CCBUG: 219873

svn path=/trunk/KDE/kdebase/workspace/; revision=1170157