Add Description to each incidence
I went ahead and added description to the list of entries which the
dataengine exported. I had considered exporting richDescription,
descriptionIsRich, etc, but it seemed like overkill when the engines
are supposed to only export the data which applets would need.
Willing to change that though.

FEATURE: Add descriptions to the calendar engine's events query.

CCMAIL: johnlayt@googlemail.com

svn path=/trunk/KDE/kdebase/workspace/; revision=1173279