Short version: cleanup code, magic numbers--
Long version:
Triggering a popup of the notifier causes the popup icon to be reset;
this implies that we when we have to both change icon and trigger a popup   
we should do it in a specific order.
This commit adds a state flag and a check to improve readability of the code
and avoid related mistakes in the future
Moreover we define two timeouts, one long and one short, in order to reduce
magic numbers appearing as timeouts in the code.

svn path=/trunk/KDE/kdebase/workspace/; revision=1189566