Change systrem tray's unhider so that is uses arrows.svg instead of systemtray.svg.
In order to allow other applets to use the arrows that are currently private to
the systray, they have been incorporated into arrows.svgz as size-hinted
versions of the generic arrows. This commit makes the systray actually use them
and also removes the arrow elements from systray.svgz as discussed on
plasma-devel.

svn path=/trunk/KDE/kdebase/workspace/; revision=1194421