Add Plasma Calendar config interface to select multiple Holiday Regions.
An earlier commit did the backend work, now here's the config part, using the
new KHolidayRegionSelector widget from KHolidays.  You can now choose to
display many different Hoiday Regions, and to choose which ones display as
a day off (i.e. in Red or labelled as Holiday in the pop-up).

Note this adds a dependency on KHolidays to the applet, but the data engines
already have that so it's nothing new.

The config widget is still a bit ugly, especially the two tickboxes thing, that
will get some polish shortly, I just feel a freeze a coming.


svn path=/trunk/KDE/kdebase/workspace/; revision=1195195