Make kdepimlibs optional requirement for kdebase again.
Turns out that kdepimlibs is only optional require elsewhere in kdebase (it was
optional at a higher level than I looked in the dataengines), so remove the
hard dependency on the new HolidayRegionSelector widget.  Instead now if there
is no kdepimlibs then there is no config option at all, which makes more sense
than an empty combo in the past.

I hope I got this right, it took a days head-scratching to get it building both
with and without kdepimlibs available.


svn path=/trunk/KDE/kdebase/workspace/; revision=1195683