Remove the backends that are broken
These backends rely on scraping the webpage and everytime the
services change the page they get broken. The sanest way to work
with this is to remove them from KDE release and start releasing them
on GHNS where we will be able to provide a fix as soon as the
breaks.


svn path=/branches/KDE/4.6/kdebase/workspace/; revision=1208950