Add support for GET protocol
Now the dataengine supports the "GET" protocol. The way to use it is to
define a function in the main script file of the plugin as follows (in JS):

function method() {
    return "GET";
}

One may also return "POST" but this is optional and also the default method.
Please, note that it is not necessary to have the above method. If it is not found
the dataengine will fallback to the default method.


svn path=/trunk/KDE/kdebase/workspace/; revision=1211568