Unbind shader before copying pixels to back buffer.
This solves the regression with NVIDIA drivers resulting in a black screen.
Apparently NVIDIA does not support glXCopySubBuffer and the copying of pixels
fails with a shader bound.

So unbinding the shader before copying the pixels and rebinding the shader
afterwards resolves the issue.

This allows to remove the temporary hack to make NVIDIA work, again.