Remove code that writes 4.7\libraryPath into ~/.config/Trolltech.conf
This was meant to make qt-only apps find kde plugins like oxygen, in a
kde desktop, but this is achieved by having startkde export QT_PLUGIN_PATH
anyway, these days. And the problem with writing to Trolltech.conf is that
it breaks having two versions of KDE on top of the same Qt. A BIC in
liboxygensyle.so and your kde apps crash on startup, due to loading the
wrong oxygen style. Discussed with RichMoore and Thiago on IRC.