put the wallpaper listing into a thread on its own
upside: application start no longer blocks on this listing. and
without a GUI to update it's rather fast. in ~500ms i'm able to
load 13000+ images from 260 dirs.

downside: if the listing is going on when the application exits,
a crash occurs because QDir and its internal file engine are no
longer happy.